transitioned season countdown board over to the new self contained board style
