--- src/boards/builtins/clock/__init__.py
@@ -1,17 +1,0 @@
-"""
-Clock board plugin demonstrating the plugin system.
-
-This plugin shows how to create a simple board that displays text and uses configuration.
-"""
-
-# Plugin metadata
-PLUGIN_NAME = "clock_plugin"
-PLUGIN_VERSION = "1.0.0"
-PLUGIN_DESCRIPTION = "Example board plugin for demonstration"
-PLUGIN_AUTHOR = "NHL LED Scoreboard"
-
-# Plugin requirements (optional)
-PLUGIN_REQUIREMENTS = []
-
-# Minimum application version required (optional)
-MIN_APP_VERSION = "1.0.0"
