--- src/boards/builtins/clock/layout_128x64.json
@@ -1,36 +1,0 @@
-{
-    "_default": {
-        "position": [0, 0],
-        "align": "left-top"
-    },
-    "clock": {
-        "message": {
-            "position": ["50%", 16],
-            "align": "center-top",
-            "font": "small"
-        },
-        "time": {
-            "position": ["50%", 28],
-            "align": "center-top",
-            "font": "default_medium"
-        },
-        "date": {
-            "position": ["50%", 40],
-            "align": "center-top",
-            "font": "small"
-        },
-        "plugin_info": {
-            "position": ["50%", 52],
-            "align": "center-top",
-            "font": "small"
-        },
-        "meridiem": {
-            "position": [2, 0],
-            "align": "left-top",
-            "relative": {
-                "to": "time",
-                "align": "right-top"
-            } 
-        }
-    }
-}
