+++ src/boards/builtins/season_countdown/__init__.py
@@ -1,0 +1,17 @@
+"""
+Season Countdown board module for NHL season countdown.
+
+This board module displays countdown to the next NHL season with customizable display.
+"""
+
+# Board metadata using standard Python package conventions
+__version__ = "1.0.0"
+__description__ = "Season Countdown Board displays time until NHL season starts"
+__board_name__ = "Season Countdown Board"
+__author__ = "NHL LED Scoreboard"
+
+# Board requirements (optional)
+__requirements__ = []
+
+# Minimum application version required (optional)
+__min_app_version__ = "1.0.0"
